Добавлена логика пересланных сообщений на бекенд

+++ MessageStorageServiceVT/src/main/java/com/alexgls/springboot/messagestorageservicevt/dto/messages/CreateMessagePayload.java
@@ -5,7 +5,7 @@ import com.alexgls.springboot.messagestorageservicevt.dto.attachments.CreateAtta
 import java.util.List;
 
 public record CreateMessagePayload(
-        Integer chatId,
+        long chatId,
         int senderId,
         String content,
         List<CreateAttachmentPayload> attachments,
